feat: selected cards are persistent across pages
